+++ Service/IService/IScoreService.cs
@@ -10,9 +10,9 @@ namespace Service.IService
     public interface IScoreService
     {
         Task<List<Score>> GetScore();
-        Task<Score> GetScoreByID(int scoreId);
+        Task<Score> GetScoreByID(Int64 scoreId);
         Task<bool> EditScore(Score scoreId);
         Task<bool> InsertScore(Score insertscore);
-        Task<bool> DeleteScore(int scoreId);
+        Task<bool> DeleteScore(Int64 scoreId);
     }
 }
